+++ BaseLib/Utilities/interfaceQueryable.hpp
